requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
google-genai>=0.2.0
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
